new uchar

+++ src/std/stdlib/constants.rs
@@ -1,3 +1,1 @@
-use crate::c_int;
-
-pub const MB_LEN_MAX: c_int = 16;
+pub const MB_LEN_MAX: usize = 16;
